+++ react/Installments.tsx
@@ -6,7 +6,7 @@ import { useProduct } from 'vtex.product-context'
 import InstallmentsRenderer, {
   CSS_HANDLES,
 } from './components/InstallmentsRenderer'
-import { getFirstAvailableSeller } from './modules/seller'
+import { getDefaultSeller } from './modules/seller'
 
 const messages = defineMessages({
   title: {
@@ -34,11 +34,9 @@ function Installments({
 }: Props) {
   const productContextValue = useProduct()
   const { handles } = useCssHandles(CSS_HANDLES, { classes })
-  const availableSeller = getFirstAvailableSeller(
-    productContextValue?.selectedItem?.sellers
-  )
+  const seller = getDefaultSeller(productContextValue?.selectedItem?.sellers)
 
-  const commercialOffer = availableSeller?.commertialOffer
+  const commercialOffer = seller?.commertialOffer
 
   if (
     !commercialOffer?.Installments ||