+++ react/ListPrice.tsx
@@ -5,7 +5,7 @@ import { FormattedCurrency } from 'vtex.format-currency'
 import { useCssHandles, CssHandlesTypes } from 'vtex.css-handles'
 import { IOMessageWithMarkers } from 'vtex.native-types'
 
-import { getFirstAvailableSeller } from './modules/seller'
+import { getDefaultSeller } from './modules/seller'
 
 const CSS_HANDLES = [
   'listPrice',
@@ -45,11 +45,9 @@ function ListPrice({
   const { handles } = useCssHandles(CSS_HANDLES, { classes })
   const productContextValue = useProduct()
 
-  const availableSeller = getFirstAvailableSeller(
-    productContextValue?.selectedItem?.sellers
-  )
+  const seller = getDefaultSeller(productContextValue?.selectedItem?.sellers)
 
-  const commercialOffer = availableSeller?.commertialOffer
+  const commercialOffer = seller?.commertialOffer
 
   if (!commercialOffer || commercialOffer?.AvailableQuantity <= 0) {
     return null